+++ include/Shape.h
@@ -5,7 +5,6 @@
 
 #include <string>
 #include <list>
-#include <cstddef>  // For size_t
 #include "ShapeType.h"
 
 template <class T> class Coord;
@@ -17,7 +16,7 @@ class Shape {
         std::string _name;
         ShapeType _type;
         std::list<Coord<double>*> _wCoords;
-        std::list<const Coord<size_t>*> _vpCoords;
+        std::list<const Coord<int>*> _vpCoords;
 
     protected:
         Shape(const std::string name, const ShapeType type);
@@ -27,9 +26,9 @@ class Shape {
         std::string get_name() const {return _name;}
         ShapeType get_type() const {return _type;}
         const std::list<Coord<double>*>* getWindowCoordinates() const {return &_wCoords;}
-        const std::list<const Coord<size_t>*>* getViewportCoordinates() const {return &_vpCoords;}
+        const std::list<const Coord<int>*>* getViewportCoordinates() const {return &_vpCoords;}
         void clearViewportCoordinates();
-        void addViewportCoordinate(const Coord<size_t> *coord);
+        void addViewportCoordinate(const Coord<int> *coord);
         virtual void clipToWindow(WorldWindow *w) = 0;
 };
